simplification + delete name

+++ include/selection.hpp
@@ -22,7 +22,7 @@ Population selection_eugenisme(int q, const Population & pop);
 
 Population selection_elitiste(int q, const Population & popParent, const Population& popEnfant);
 
-double somme_dist_individus(const Population & pop);
+float somme_dist_individus(const Population & pop);
 vector<int> liste_triee_individus(const Population & pop);
 
 Population selection(Choix choix,int q, const Population & pop);
